requirements.txt
# Requires
pytorch
numpy
matplotlib
